Add a watchdog feature (#324)
For LogSourceFactories that implement it, a Watchdog function can be used
to signal that the log source is unhealthy.

For instance, for kubernetes, the configured pod replicas may have changed,
and the log source should be reloaded to track the changed replicas.

In this simple initial implementation, the main function simply exits when the watchdog return unhealthy.
In a more ambitious future improvement, it can instead close and re-initialize the log sources and exporters.
Currently, only the kubernetes log source implements Watchdog(), so it is fine to let the main program exit,
as kubernetes can take care of respawning the postfix_exporter pods.